Improve ShoppingList func, Flows, etc

+++ app/src/main/java/com/alexmncn/flemingshop/data/repository/ShoppingListRepository.kt
@@ -1,5 +1,6 @@
 package com.alexmncn.flemingshop.data.repository
 
+import android.util.Log
 import com.alexmncn.flemingshop.data.db.AppDatabase
 import com.alexmncn.flemingshop.data.db.ArticleItem
 import kotlinx.coroutines.flow.Flow
@@ -27,19 +28,23 @@ class ShoppingListRepository(db: AppDatabase) {
     suspend fun insertArticle(articleItem: ArticleItem) {
         val existingArticle = articleItemDao.getByCodebar(articleItem.codebar)
 
-        if (existingArticle == null) {
-            // Si no existe, lo insertamos
+        Log.d("existingArticle", existingArticle.toString())
+
+        if (existingArticle == null && articleItem.quantity > 0) {
+            // Si no existe y la cantidad es positiva, lo insertamos
             articleItemDao.insert(articleItem)
         } else { // Si ya existe
             // Calculamos la nueva cantidad
-            val newQuantity = existingArticle.quantity + articleItem.quantity
+            val newQuantity = existingArticle?.quantity?.plus(articleItem.quantity)
 
-            if (newQuantity > 0) {
-                // Si es mayor que 0, lo actualizamos
-                articleItemDao.updateQuantity(articleItem.codebar, newQuantity)
-            } else {
-                // Sino, lo eliminamos
-                articleItemDao.deleteByCodebar(articleItem.codebar)
+            if (newQuantity != null) {
+                if (newQuantity > 0) {
+                    // Si la nueva cantidad es positiva, lo actualizamos
+                    articleItemDao.updateQuantity(articleItem.codebar, newQuantity)
+                } else {
+                    // Sino, lo eliminamos
+                    articleItemDao.deleteByCodebar(articleItem.codebar)
+                }
             }
         }
     }
